Simple 2D 1-body trajectory optimizer.
This is my first attempt at making a trajectory optimizer. It will cycle through smaller and smaller random variations in each direction to find the optimal starting path to get as close to a target point as possible. The second part of the code will just display the trajectory of the most optimal path in a visual way. It's just a test really, so the code is quite messy and it isn't incredibly easy to use. To change the "hyperparameters", the randrange changes the magnitude of the randomly selected velocities for the simulation, the step size dictates how often physics will be calculated, and the maxsimulationtime will end the simulation after a certain time. The optimizer will try to reach your designated point at the end of the simulation time.

The "tweaking" is just every stage where the random range of velocities is lowered to become more precise.